Add ContainerSize setting feature

+++ app/src/main/java/com/example/hydrationappcompose/data/repository/SettingsRepository.kt
@@ -3,11 +3,11 @@ package com.example.hydrationappcompose.data.repository
 interface SettingsRepository {
     suspend fun saveUnitId(unitId: Int)
 
-    suspend fun saveDailyGoal(dailyGoal: Int)
+    suspend fun saveDailyGoal(dailyGoal: Float)
 
-    suspend fun saveContainer1Size(size: Int)
+    suspend fun saveContainer1Size(size: Float)
 
-    suspend fun saveContainer2Size(size: Int)
+    suspend fun saveContainer2Size(size: Float)
 
-    suspend fun saveContainer3Size(size: Int)
+    suspend fun saveContainer3Size(size: Float)
 }
