+++ app/src/main/java/com/example/hydrationappcompose/data/repository/SettingsRepositoryImpl.kt
@@ -5,11 +5,11 @@ import com.example.hydrationappcompose.data.datasource.SettingsDataSource
 class SettingsRepositoryImpl(private val dataSource: SettingsDataSource) : SettingsRepository {
     override suspend fun saveUnitId(unitId: Int) = dataSource.saveUnitId(unitId)
 
-    override suspend fun saveDailyGoal(dailyGoal: Int) = dataSource.saveDailyGoal(dailyGoal)
+    override suspend fun saveDailyGoal(dailyGoal: Float) = dataSource.saveDailyGoal(dailyGoal)
 
-    override suspend fun saveContainer1Size(size: Int) = dataSource.saveContainer1Size(size)
+    override suspend fun saveContainer1Size(size: Float) = dataSource.saveContainer1Size(size)
 
-    override suspend fun saveContainer2Size(size: Int) = dataSource.saveContainer2Size(size)
+    override suspend fun saveContainer2Size(size: Float) = dataSource.saveContainer2Size(size)
 
-    override suspend fun saveContainer3Size(size: Int) = dataSource.saveContainer3Size(size)
+    override suspend fun saveContainer3Size(size: Float) = dataSource.saveContainer3Size(size)
 }
