+++ app/src/main/java/com/example/hydrationappcompose/domain/model/Settings.kt
@@ -2,8 +2,8 @@ package com.example.hydrationappcompose.domain.model
 
 data class Settings(
     val unitId: Int,
-    val dailyGoal: Int,
-    val container1Size: Int,
-    val container2Size: Int,
-    val container3Size: Int
+    val dailyGoal: Float,
+    val container1Size: Float,
+    val container2Size: Float,
+    val container3Size: Float
 )
